+++ apps/extension/src/ui/apps/dashboard/routes/AccountAdd/AccountAddDerivedPage.tsx
@@ -1,7 +1,7 @@
 import { AccountAddressType } from "@core/domains/accounts/types"
 import { HeaderBlock } from "@talisman/components/HeaderBlock"
 import { Spacer } from "@talisman/components/Spacer"
-import { AccountAddDerivedForm } from "@ui/domains/Account/AccountCreate/AccountAddDerivedForm"
+import { AccountAddDerivedForm } from "@ui/domains/Account/AccountAdd/AccountAddDerivedForm"
 import { useSelectAccountAndNavigate } from "@ui/hooks/useSelectAccountAndNavigate"
 import { useTranslation } from "react-i18next"
 import { useSearchParams } from "react-router-dom"
