+++ apps/extension/src/ui/apps/dashboard/routes/AccountAdd/AccountAddSecretWizard/index.tsx
@@ -1,4 +1,4 @@
-import { AccountAddSecretWizard } from "@ui/domains/Account/AccountCreate/AccountAddSecret/router"
+import { AccountAddSecretWizard } from "@ui/domains/Account/AccountAdd/AccountAddSecret/router"
 import { useSelectAccountAndNavigate } from "@ui/hooks/useSelectAccountAndNavigate"
 
 import { AccountAddSecretLayout } from "./AccountAddSecretLayout"
