+++ apps/extension/src/ui/apps/onboard/routes/AddAccount/AccountAddDerivedPage.tsx
@@ -1,7 +1,7 @@
 import { OnboardDialog } from "@ui/apps/onboard/components/OnboardDialog"
 import { useOnboard } from "@ui/apps/onboard/context"
 import { Layout } from "@ui/apps/onboard/layout"
-import { AccountAddDerivedForm } from "@ui/domains/Account/AccountCreate/AccountAddDerivedForm"
+import { AccountAddDerivedForm } from "@ui/domains/Account/AccountAdd/AccountAddDerivedForm"
 import { useTranslation } from "react-i18next"
 
 export const AccountAddDerivedPage = () => {
