+++ apps/extension/src/ui/apps/onboard/routes/AddAccount/AccountAddSecretWizard/index.tsx
@@ -1,5 +1,5 @@
 import { useOnboard } from "@ui/apps/onboard/context"
-import { AccountAddSecretWizard } from "@ui/domains/Account/AccountCreate/AccountAddSecret/router"
+import { AccountAddSecretWizard } from "@ui/domains/Account/AccountAdd/AccountAddSecret/router"
 
 import { AccountAddSecretLayout } from "./AccountAddSecretLayout"
