+++ apps/extension/src/ui/apps/onboard/routes/AddAccount/index.tsx
@@ -2,7 +2,7 @@ import { AnalyticsPage, sendAnalyticsEvent } from "@ui/api/analytics"
 import { OnboardDialog } from "@ui/apps/onboard/components/OnboardDialog"
 import { useOnboard } from "@ui/apps/onboard/context"
 import { Layout } from "@ui/apps/onboard/layout"
-import { AccountCreateMenu } from "@ui/domains/Account/AccountCreate"
+import { AccountCreateMenu } from "@ui/domains/Account/AccountAdd"
 import { useAnalyticsPageView } from "@ui/hooks/useAnalyticsPageView"
 import { useCallback } from "react"
 import { useTranslation } from "react-i18next"
