+++ apps/extension/src/ui/domains/Account/AccountAdd/AccountAddSecret/router.tsx
